makepkg: implement alternative package splitting approach
The original package splitting approach in makepkg was aligned to the KDE
packaging approach (at the time) where each component had its own
"make install-foo" target.  In practice, distributions want to split a wider
variety of packages that do not fit this mould, and various hacks have arisen.

Implement a secondary approach to package splitting.  A PKGBUILD can provide
a 'stash()' function that is run between building and packaging.  This
function will typically do the full install of the built software into the
$stashdir directory.  I.e. run a command similar to
"make DESTDIR=$stashdir install".  Each 'package_foo()' function can then
include a list of files that are transferred from the stash into the package,
in addition to any other packaging commands.

For example:
filelist=('usr/bin/pacman'
          'usr/lib/*')

The filelist accepts glob patterns as understood by bash with extglob
enabled - including '*', '?' and '[]'.  Files can be excluded from the
filelist by prefixing with the '!' character.

Files are removed from the stash as they are included in packages.  This
allows makepkg to warn about any files that are not packaged.

Signed-off-by: Allan McRae <allan@archlinux.org>